requirements.txt
qdrant-client==0.0.1
python-dotenv==1.0.0
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
